+++ packages/toolbars-new/web/src/Toolbar/buttonsListCreatorConsts.tsx
@@ -39,6 +39,10 @@ import {
   UndoIcon,
   RedoIcon,
   PlusIcon,
+  AlignTextCenterIcon,
+  AlignJustifyIcon,
+  AlignLeftIcon,
+  AlignRightIcon,
 } from '../icons';
 import LinkModal from '../modals/link/LinkComponents/LinkModal';
 import AlignmentPanel from '../modals/alignment/AlignmentPanel';
@@ -54,6 +58,40 @@ export const HEADING_TYPE_TO_ELEMENT = Object.freeze({
   'header-six': 'H6',
   unstyled: 'P',
 });
+
+export const alignmentsModalData = [
+  {
+    text: 'AlignObject_Left_Tooltip',
+    tooltip: 'AlignObject_Left_Tooltip',
+    commandKey: 'left',
+    icon: AlignLeftIcon,
+  },
+  {
+    text: 'AlignObject_Center_Tooltip',
+    tooltip: 'AlignObject_Center_Tooltip',
+    commandKey: 'center',
+    icon: AlignTextCenterIcon,
+  },
+  {
+    text: 'AlignObject_Right_Tooltip',
+    tooltip: 'AlignObject_Right_Tooltip',
+    commandKey: 'right',
+    icon: AlignRightIcon,
+  },
+  {
+    text: 'AlignTextJustifyButton_Tooltip',
+    tooltip: 'AlignTextJustifyButton_Tooltip',
+    commandKey: 'justify',
+    icon: AlignJustifyIcon,
+  },
+];
+
+export const lineSpacingModalData = [
+  { text: '1', commandKey: '1' },
+  { text: '2', commandKey: '2' },
+  { text: '2.5', commandKey: '2.5' },
+  { text: '3', commandKey: '3' },
+];
 
 type buttonsFullDataType = {
   type: string;